backend_sansar/app/routers/ai_router.py
# routers/ai_router.py
from fastapi import APIRouter, Depends
import cv2
import torch  # Assume model for presence detection

ai_router = APIRouter(prefix="/ai", tags=["ai"])

@ai_router.post("/analyze-presence/{device_id}")
async def analyze_presence(device_id: int):
    # Fetch stream, analyze frame
    # Stub: use cv2 to capture, torch model to detect
    return {"presence": True}