backend_sansar/requirements.txt
# requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy[asyncio]==2.0.0
asyncpg==0.29.0
alembic==1.9.0
fastapi-users[sqlalchemy,jwt]==12.1.2
pyotp==2.8.0
celery==5.2.7
redis==4.5.1
twilio==7.17.0
python-telegram-bot==20.1
pandas==1.5.3
xlsxwriter==3.0.9
ping3==4.0.4
requests==2.28.2
pytest==7.2.2
httpx==0.23.3
#opencv-python==4.8.0.74  # For AI analysis if needed
#torch==2.4.1